# :fire: add config to enable splitting by comma in parsers