## Description

This PR adds a config field to enable the feature that was implemented in https://github.com/gofiber/fiber/pull/817. It's disabled by default since it's not common pattern in web frameworks. Also added escape char support to avoid splitting some fields when we want.
Fixes #2557 

Benchmarks Before:
```
Benchmark_Ctx_BodyParser_JSON
Benchmark_Ctx_BodyParser_JSON-16                                                     	 2245416	       533.9 ns/op	     240 B/op	       6 allocs/op
Benchmark_Ctx_BodyParser_XML
Benchmark_Ctx_BodyParser_XML-16                                                      	  626001	      1902 ns/op	    1160 B/op	      24 allocs/op
Benchmark_Ctx_BodyParser_Form
Benchmark_Ctx_BodyParser_Form-16                                                     	 1261950	       953.7 ns/op	     360 B/op	      13 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm
Benchmark_Ctx_BodyParser_MultipartForm-16                                            	 1412667	       836.6 ns/op	     328 B/op	      12 allocs/op
Benchmark_Ctx_QueryParser
Benchmark_Ctx_QueryParser-16                                                         	  461858	      2507 ns/op	     857 B/op	      38 allocs/op
Benchmark_Ctx_parseQuery
Benchmark_Ctx_parseQuery-16                                                          	  550956	      2174 ns/op	     753 B/op	      29 allocs/op
Benchmark_Ctx_QueryParser_Comma
Benchmark_Ctx_QueryParser_Comma-16                                                   	  424041	      2798 ns/op	     929 B/op	      44 allocs/op
Benchmark_Ctx_ReqHeaderParser
Benchmark_Ctx_ReqHeaderParser-16                                                     	  433765	      2757 ns/op	     929 B/op	      44 allocs/op
```
Benchmarks After:
```
Benchmark_Ctx_Body_With_Compression
Benchmark_Ctx_Body_With_Compression-16                                               	 2044170	       588.4 ns/op
Benchmark_Ctx_BodyParser_JSON
Benchmark_Ctx_BodyParser_JSON-16                                                     	 2250416	       536.8 ns/op	     240 B/op	       6 allocs/op
Benchmark_Ctx_BodyParser_XML
Benchmark_Ctx_BodyParser_XML-16                                                      	  618874	      1897 ns/op	    1160 B/op	      24 allocs/op
Benchmark_Ctx_BodyParser_Form
Benchmark_Ctx_BodyParser_Form-16                                                     	 1239142	       968.6 ns/op	     360 B/op	      13 allocs/op
Benchmark_Ctx_BodyParser_MultipartForm
Benchmark_Ctx_BodyParser_MultipartForm-16                                            	 1432702	       851.8 ns/op	     328 B/op	      12 allocs/op
Benchmark_Ctx_QueryParser
Benchmark_Ctx_QueryParser-16                                                         	  459550	      2566 ns/op	     856 B/op	      38 allocs/op
Benchmark_Ctx_parseQuery
Benchmark_Ctx_parseQuery-16                                                          	  538406	      2205 ns/op	     753 B/op	      29 allocs/op
Benchmark_Ctx_QueryParser_Comma
Benchmark_Ctx_QueryParser_Comma-16                                                   	  403694	      2885 ns/op	     945 B/op	      45 allocs/op
Benchmark_Ctx_ReqHeaderParser
Benchmark_Ctx_ReqHeaderParser-16                                                     	  403273	      2857 ns/op	     945 B/op	      45 allocs/op
```
## Type of change
- [x] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] This change requires a documentation update

## Checklist:

- [x] For new functionalities I follow the inspiration of the express js framework and built them similar in usage
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation - /docs/ directory for https://docs.gofiber.io/
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] I tried to make my code as fast as possible with as few allocations as possible

## Commit formatting:
Use emojis on commit messages so it provides an easy way of identifying the purpose or intention of a commit. Check out the emoji cheatsheet here: https://gitmoji.carloscuesta.me/